requirements.txt
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
distlib==0.3.6
filelock==3.11.0
google-api-core==2.11.0
google-api-python-client==2.85.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
novas==3.1.1.5
package-name==0.1
platformdirs==3.2.0
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dotenv==1.0.0
requests==2.28.2
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.21.0

pip~=23.1.1
wheel~=0.40.0
Pillow~=9.4.0
setuptools~=67.6.1
isodate~=0.6.1